LayoutTests/fast/canvas/canvas-resize-reset-pixelRatio-expected.txt

Testing webkitBackingStorePixelRatio after resize
PASS: webkitBackingStorePixelRatio is 1.
PASS: webkitBackingStorePixelRatio is 2.
PASS: webkitBackingStorePixelRatio is 1.
PASS: webkitBackingStorePixelRatio is 2.
